Added regexp matching for type alias keys
It's a bit less efficient now that it has to loop through keys each time. I could add a cache but that seems like extra work. Eventually, if performance profiling shows that to be a bad thing, that can happen.

Splits out some pieces of `fillOutRawOptions` into separate files because the one file was getting too big.

Fixes #9.