Moved custom/built-in mutation resolving earlier, to option parsing
As @ian-craig noted, it's a little odd to choose between these things within `findMutationsInFile`. The "which mutations will we apply?" logic should really be part of options parsing. The work to look at the raw options for them is already done in `options/addedMutators.ts`, so this tweaks the logic to default to `builtInFileMutators` instead of an empty array.

Starts on #96.