Added auto-installation of missing types
Fixes #102.

Adds a whole new section of configuration, `package`. This contains `directory` and/or `file` and/or `missingTypes` settings. The directory one is interesting because now other things will be relative to it. These are necessary in the context of the change because the package file needs to be read from and a `yarn.lock` in the directory is necessary to auto-detect Yarn.

Wowee would it be nice to get the test harness able to support cross-file changes and resets...