Exposed and used TypeScript's internal isTypeAssignableTo for type comparisons
Removes the clunky homegrown type checking logic and replaces it with the method within the type checker. The method is exposed on the tc by manually rewriting TypeScript's `tsc.js`. Very dirty.

My next commit will make it check the first few hundred chars of the file for a `/* TypeStat! */` comment so it doesn't have to re-read that whole file constantly.