test: use vitest for mutation tests (#1476)
<!-- 👋 Hi, thanks for sending a PR to TypeStat! 💖.
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #1454
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/TypeStat/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/TypeStat/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

<!-- Description of what is changed and how the code change does that.
-->

This uses vitest instead of mocha to run mutation test. It also skips
using `automutate-test` and just writes output straight to file and
compares that. This allows us to get code coverage!

However, there are now some output changes with tests. I will comment on
them to highlight those.

Also, I will cleanup some unrelated changes.

Emoji: 😹 

---

`runMutationForTest` is basically copy of code in `testSetup.ts`. What
it does is, that it copies `original.ts` file to file named `actual.ts`.
Then it runs mutations in that file and compares the result to
`expected.ts` file using vitest's file snapshot feature.

With this approach, `typestat.json`does not need to be it's own file but
we could just pass the settings as value from test case to automutate.
I'm wondering, could this be possible also for tsconfig. However, I did
not touch this logic now. This could be later improvement.

One difference between old and new approach is, that `waves: { maximum:
infiniteWaveThreshold + 1 }` this settings is not used for new approach.
I'm not sure if that is needed, and what it was doing underneath. It
seems that all the test work similarly without this.

---------

Co-authored-by: Josh Goldberg ✨ <git@joshuakgoldberg.com>