fix: simplify collecting strict TS options (#2196)
## PR Checklist

- [ ] Addresses an existing open issue: fixes #000 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/TypeStat/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/TypeStat/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

I forgot this from #2192 , though maybe it's better to have it
separately.

This simplifies and IMO corrects the logic for some of these options.
Before, it was possible that if you have `noImplicitAny` set as false in
`typestat.json` file for the fixes, it would change the typescript
compiler options to be also false. This could lead to weird behavior.

New logic set `noImplicitAny` as true in the collected compiler options
if it's true either in the fixes section or in the original ts config
file.

I think the `types` section in `BaseTypeStatOptions` is unnecessary.
It's only used in `findComplaintForOptions` to check options and print
"fixes.strictNonNullAssertions specified but not strictNullChecks." if
there is mismatch. That check could use the actual collected compiler
options. However, it's not harmful to keep it.